% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_gov_api_key}
\alias{data_gov_api_key}
\title{Save a data.gov API key}
\usage{
data_gov_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{install}{}
}
\description{
It's best practice not to include your API key in scripts. Save it interactively with this function, then it will be available with Sys.getenv("DATAGOV_API_KEY"). All 'api_key' arguments in tidyfec functions default to Sys.getenv("DATAGOV_API_KEY").
}
