% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{get_candidate_totals}
\alias{get_candidate_totals}
\title{Search FEC records for Candidates}
\usage{
get_candidate_totals(data_structure = "both", candidate_ids = NULL,
  sort_null_only = NULL, cycle = NULL, sort = NULL, designation = NULL,
  type = NULL, full_election = TRUE, api_key = NULL)
}
\arguments{
\item{data_structure}{A character describing how you want OpenFEC results. Options are 'tidy', 'list', or 'both.' Choose list for raw results.}

\item{candidate_ids}{A character vector of candidate ids to get financial totals for. Required to run.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null.}

\item{cycle}{Two-year election cycle in which a candidate runs for office. Calculated from FEC Form 2. The cycle begins with an odd year and is named for its ending, even year. This cycle follows the traditional house election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. To see data for the entire four years of a presidential term or six years of a senatorial term, you will need the election_full flag.}

\item{sort}{Provide a field to sort by. Use - for descending order.}

\item{designation}{The one-letter designation code of the organization: - A: authorized by a candidate - J: joint fundraising committee - P: principal campaign committee of a candidate - U: unauthorized - B: lobbyist/registrant PAC - D: leadership PAC}

\item{type}{The one-letter type code of the organization:  - C: communication cost - D: delegate - E: electioneering communication - H: House - I: independent expenditor (person or group) - N: PAC - nonqualified - O: independent expenditure-only (super PACs) - P: presidential - Q: PAC - qualified - S: Senate - U: single candidate independent expenditure - V: PAC with non-contribution account, nonqualified - W: PAC with non-contribution account, qualified - X: party, nonqualified - Y: party, qualified - Z: national party non-federal account}

\item{full_election}{Get totals for full election period. Boolean.}

\item{api_key}{API key for https://api.data.gov. Get one at https://api.data.gov/signup.}
}
\description{
tidycensus Documentation: Search for candidates. If tidy is TRUE, each row describes a candidate during a specific election cycle. The same candidate may have rows for multiple elections depending on search paramters.
}
\details{
OpenFEC Documentation: Fetch basic information about candidates and their principal committees.

Each result reflects a unique FEC candidate ID. That ID is assigned to the candidate for a particular office sought. If a candidate runs for the same office over time, that ID stays the same. If the same person runs for multiple offices — for example, a House candidate runs for a Senate office — that candidate will get a unique ID for each office.

The candidate endpoints primarily use data from FEC registration Form 1, for candidate information, and Form 2, for committees information, with additional information to provide context.
}
