% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_election_summary.R
\name{get_election_summary}
\alias{get_election_summary}
\title{Get Election Summary}
\usage{
get_election_summary(data = NULL, data_structure = "tidy",
  api_key = Sys.getenv("DATAGOV_API_KEY"), district = NULL, cycle = NULL,
  office = NULL, election_full = NULL, state = NULL)
}
\arguments{
\item{data}{A dataframe or tibble. Usually this will be the returned result of search_elections(). If columns called cycle, office, district, and state are in the table, . Either this argument or candidate_id is required.}

\item{data_structure}{A character describing how you want OpenFEC results. Options are 'tidy', 'list', or 'both.' Choose list for raw results.}

\item{api_key}{API key for https://api.data.gov. Get one at https://api.data.gov/signup.}

\item{district}{Two-digit US House distirict of the office the candidate is running for. Presidential, Senate and House at-large candidates will have District 00.}

\item{cycle}{Two-year election cycle in which a candidate runs for office. Calculated from FEC Form 2. The cycle begins with an odd year and is named for its ending, even year. This cycle follows the traditional house election cycle and subdivides the presidential and Senate elections into comparable two-year blocks. To see data for the entire four years of a presidential term or six years of a senatorial term, you will need the election_full flag.}

\item{office}{Federal office candidate runs for: H, S or P}

\item{election_full}{Aggregate values over full election period}

\item{state}{US state or territory where a candidate runs for office}
}
\description{
Get Election Summary
}
\examples{

}
