% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_candidate_committees.R
\name{unnest_candidate_committees}
\alias{unnest_candidate_committees}
\title{Unnest Candidate Committees}
\usage{
unnest_candidate_committees(data)
}
\arguments{
\item{data}{}
}
\description{
Unnest Candidate Committees
}
