% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_principal_committees.R
\name{unnest_principal_committees}
\alias{unnest_principal_committees}
\title{Unnest the principal committees associated with candidates}
\usage{
unnest_principal_committees(data = NULL)
}
\arguments{
\item{data}{}
}
\description{
This function prepares the results of search_candidates for get_ functions that depend on committee_id.
}
\details{
For results of search_candidates, each row is an individual candidate. The principal_committees list column contains nested information for principal committees associated with candidates. Returns a tibble with each row describing a unique candidate and principal committee pair. Some candidates will have no principal committee (maybe they just got in, didn't do well, or didn't raise much) and others will have multiple principal committees (they formed new committees for new elections).
}
