% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee_reports.R
\name{get_committee_reports}
\alias{get_committee_reports}
\title{Get Committee Reports}
\usage{
get_committee_reports(data = NULL,
  api_key = Sys.getenv("DATAGOV_API_KEY"), data_structure = "tidy",
  committee_id = NULL, sort_null_only = NULL,
  min_cash_on_hand_end_period_amount = NULL,
  beginning_image_number = NULL, report_type = NULL, year = NULL,
  sort_hide_null = NULL, max_cash_on_hand_end_period_amount = NULL,
  min_independent_expenditures = NULL, min_debts_owed_amount = NULL,
  max_total_contributions = NULL, type = NULL, is_amended = NULL,
  max_independent_expenditures = NULL, cycle = NULL,
  candidate_id = NULL, min_disbursements_amount = NULL,
  min_receipts_amount = NULL, max_receipts_amount = NULL,
  sort = NULL, max_disbursements_amount = NULL,
  min_party_coordinated_expenditures = NULL,
  max_debts_owed_expenditures = NULL,
  max_party_coordinated_expenditures = NULL,
  min_total_contributions = NULL)
}
\arguments{
\item{data}{A dataframe or tibble. Usually this will be the returned result of search_candidates(). If a column is called 'committee_id', get_itemized_contributions() will return results for all IDs in that column and will attempt to join data to the result by committee_id.}

\item{api_key}{api_key API key for https://api.data.gov. Get one at https://api.data.gov/signup.}

\item{data_structure}{A character describing how you want OpenFEC results. Options are 'tidy', 'list', or 'both.' Choose list for raw results.}

\item{committee_id}{A unique identifier assigned to each committee or filer registered with the FEC. In general committee id's begin with the letter C which is followed by eight digits.}

\item{sort_null_only}{Toggle that filters out all rows having sort column that is non-null}

\item{min_cash_on_hand_end_period_amount}{Filter for all amounts greater than a value.}

\item{beginning_image_number}{Unique identifier for the electronic or paper report. This number is used to construct PDF URLs to the original document.}

\item{report_type}{Report type; prefix with "-" to exclude. Name of report where the underlying data comes from:

- 10D Pre-Election
- 10G Pre-General
- 10P Pre-Primary
- 10R Pre-Run-Off
- 10S Pre-Special
- 12C Pre-Convention
- 12G Pre-General
- 12P Pre-Primary
- 12R Pre-Run-Off
- 12S Pre-Special
- 30D Post-Election
- 30G Post-General
- 30P Post-Primary
- 30R Post-Run-Off
- 30S Post-Special
- 60D Post-Convention
- M1  January Monthly
- M10 October Monthly
- M11 November Monthly
- M12 December Monthly
- M2  February Monthly
- M3  March Monthly
- M4  April Monthly
- M5  May Monthly
- M6  June Monthly
- M7  July Monthly
- M8  August Monthly
- M9  September Monthly
- MY  Mid-Year Report
- Q1  April Quarterly
- Q2  July Quarterly
- Q3  October Quarterly
- TER Termination Report
- YE  Year-End
- 90S Post Inaugural
Supplement
- 90D Post Inaugural
- 48  48 Hour Notification
- 24  24 Hour Notification
- M7S July Monthly/
 Semi-Annual
- MSA Monthly Semi-Annual
(MY)
- MYS Monthly Year End/
 Semi-Annual
- Q2S July Quarterly/
 Semi-Annual
- QSA Quarterly Semi-Annual
(MY)
- QYS Quarterly Year End/
 Semi-Annual
- QYE Quarterly Semi-Annual
(YE)
- QMS Quarterly Mid-Year/
 Semi-Annual
- MSY Monthly Semi-Annual
(YE)}

\item{year}{Year that the record applies to. Sometimes records are amended in subsequent years so this can differ from underlying form's receipt date.}

\item{sort_hide_null}{Hide null values on sorted column(s).}

\item{max_cash_on_hand_end_period_amount}{Filter for all amounts less than a value.}

\item{min_independent_expenditures}{Filter for all amounts greater than a value.}

\item{min_debts_owed_amount}{Filter for all amounts greater than a value.}

\item{max_total_contributions}{Filter for all amounts less than a value.}

\item{type}{The one-letter type code of the organization: - C communication cost
- D delegate
- E electioneering communication
- H House
- I independent expenditor (person or group)
- N PAC - nonqualified
- O independent expenditure-only (super PACs)
- P presidential
- Q PAC - qualified
- S Senate
- U single candidate independent expenditure
- V PAC with non-contribution account, nonqualified
- W PAC with non-contribution account, qualified
- X party, nonqualified
- Y party, qualified
- Z national party non-federal account}

\item{is_amended}{Report has been amended}

\item{max_independent_expenditures}{Filter for all amounts less than a value.}

\item{cycle}{Filter records to only those that were applicable to a given two-year period.The cycle begins with an odd year and is named for its ending, even year.}

\item{candidate_id}{A unique identifier assigned to each candidate registered with the FEC. If a person runs for several offices, that person will have separate candidate IDs for each office.}

\item{min_disbursements_amount}{Filter for all amounts greater than a value.}

\item{min_receipts_amount}{Filter for all amounts greater than a value.}

\item{max_receipts_amount}{Filter for all amounts less than a value.}

\item{sort}{Provide a field to sort by. Use - for descending order.}

\item{max_disbursements_amount}{Filter for all amounts less than a value.}

\item{min_party_coordinated_expenditures}{Filter for all amounts greater than a value.}

\item{max_debts_owed_expenditures}{Filter for all amounts less than a value.}

\item{max_party_coordinated_expenditures}{Filter for all amounts less than a value.}

\item{min_total_contributions}{Filter for all amounts greater than a value.}
}
\description{
Each report represents the summary information from FEC Form 3, Form 3X and Form 3P. These reports have key statistics that illuminate the financial status of a given committee. Things like cash on hand, debts owed by committee, total receipts, and total disbursements are especially helpful for understanding a committee's financial dealings.
}
\details{
By default, this endpoint includes both amended and final versions of each report. To restrict to only the final versions of each report, use is_amended=false; to view only reports that have been amended, use is_amended=true.
Several different reporting structures exist, depending on the type of organization that submits financial information. To see an example of these reporting requirements, look at the summary and detailed summary pages of FEC Form 3, Form 3X, and Form 3P.
}
