% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{search_candidates}
\alias{search_candidates}
\title{Make a request to the OpenFEC API}
\usage{
search_candidates(candidate_id = NULL, q = NULL,
  max_first_file_date = NULL)
}
\arguments{
\item{candidate_id}{A character string of the OpenFEC endpoint.}

\item{q}{A list with values of the parameters to use in the request.}
}
\description{
Fetch basic information about candidates and their principal committees.
}
\details{
Each result reflects a unique FEC candidate ID. That ID is assigned to the candidate for a particular office sought. If a candidate runs for the same office over time, that ID stays the same. If the same person runs for multiple offices — for example, a House candidate runs for a Senate office — that candidate will get a unique ID for each office.

The candidate endpoints primarily use data from FEC registration Form 1, for candidate information, and Form 2, for committees information, with additional information to provide context.
}
